@@ -5,14 +5,11 @@ using Newtonsoft.Json;
 
 namespace SourceExpander
 {
-    internal static partial class JsonUtil
+    internal static class JsonUtil
     {
-        public static JsonConverterCollection Converters { get; } = new();
-
         public static string ToJson<T>(T infos)
             => JsonConvert.SerializeObject(infos, new JsonSerializerSettings
             {
-                Converters = Converters,
                 Formatting = Formatting.None,
                 StringEscapeHandling = StringEscapeHandling.Default,
             });
@@ -22,10 +19,7 @@ namespace SourceExpander
         {
             try
             {
-                return JsonConvert.DeserializeObject<T>(json, new JsonSerializerSettings
-                {
-                    Converters = Converters,
-                })!;
+                return JsonConvert.DeserializeObject<T>(json, new JsonSerializerSettings { })!;
             }
             catch (Exception e)
             {
@@ -37,8 +31,6 @@ namespace SourceExpander
             try
             {
                 var serializer = new JsonSerializer();
-                foreach (var conv in Converters)
-                    serializer.Converters.Add(conv);
                 using var sr = new StreamReader(jsonStream);
                 using var jsonTextReader = new JsonTextReader(sr);
                 return serializer.Deserialize<T>(jsonTextReader)!;
@@ -48,11 +40,5 @@ namespace SourceExpander
                 throw new ParseJsonException(e);
             }
         }
-    }
-
-    internal sealed class ParseJsonException : Exception
-    {
-        public ParseJsonException(Exception inner) : base(inner.Message, inner)
-        { }
     }
 }
