@@ -5,6 +5,6 @@ namespace SourceExpander
 {
     internal class AssemblyUtil
     {
-        public static readonly Version AssemblyVersion = Assembly.GetExecutingAssembly().GetName().Version;
+        public static readonly Version AssemblyVersion = Assembly.GetExecutingAssembly().GetName().Version!;
     }
 }
